python3/requirements.txt
autosemver==0.5.3
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
certifi==2017.11.5
configobj==5.0.6
dulwich==0.18.6
enum34==1.1.6
idna==2.6
jieba==0.39
PyYAML==3.12
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
Whoosh==2.7.4
workflow==2.1.3
